Language and desc correction
Removed the "Terrorist can win when Exiled" and did a correction on a Cammo info that had a mayus where doesn't belong in string.csv. Also updated the README.md to have better language and removed the spelling errors. Tried to change the commands for Jackal and Trapper in ChatCommandPatch but there are so many code on it